Remove debug build condition for uvexposure reporter.
UvExposureReporter is installed on user builds but its sepolicy rules
are only active for debug builds. This change remove the debug build
condition and prevents selinux denials on the app in user builds.

The change is merged on master, cherry-picking it back to QPR as
the feature is approved in b/161648310.

Bug: 161648310
Test: Flash user build on R3 and verify no app crash.
Change-Id: I0703c0de70a69c09b493ab90be650bddc8bc8937