B5R3: Set vendor.audio.adm.buffering.ms to 2
note: As per the previous investigation on the property vendor.audio.adm.buffering.ms on its value and sepolicy on B5R3/S5, Qualcomm suggested the value change from 3 to 2. The property is to define the buffer size in ADM(Audio Data Manager) and it might lead to glitches or extra latency if it is not configured properly.

Bug: 160107932
Test: Verified with Oboe tool.
Signed-off-by: Jason Hsu <jasonchhsu@google.com>
Change-Id: Ie6184bd2170b1dbfe5bb6efa845d6994a4f9861a