Suppress the denials related to persist.vendor.service.bdroid.bdaddr
Currently redbull reads the Bluetooth Device Address from "ro.vendor.bt.bdaddr_path",
which is a CDT kernel node.
Only when redbull cannot get the BDA from CDT, it generates a random address and set
it to "persist.vendor.service.bdroid.bdaddr" as a fallback.

This denial is not what we expect to happen. It is only shown on the abnormal device.
Now suppress this denial since we do not need it.

Test: make selinux_policy
Bug: 172409472
Bug: 175310616
Change-Id: Ia1143fd1a107a8a4596726837f213935d76ca7b0
Merged-In: Ia1143fd1a107a8a4596726837f213935d76ca7b0